[2365] Added step wizard for schools
