+++ app/controllers/trainees/diversity/disclosures_controller.rb
@@ -38,13 +38,18 @@ module Trainees
       end
 
       def step_wizard
-        @step_wizard ||= Diversities::StepWizard.new(trainee: trainee)
+        @step_wizard ||= Wizards::Diversities::StepWizard.new(trainee: trainee, page_tracker: page_tracker)
       end
 
       def validate_form_completedness
         return unless trainee.diversity_disclosed?
+        return if user_came_from_backlink?
 
         redirect_to step_wizard.start_point if any_form_partially_complete?
+      end
+
+      def user_came_from_backlink?
+        session[:origin_path]&.include?("edit")
       end
 
       def any_form_partially_complete?
