+++ app/controllers/trainees/lead_schools_controller.rb
@@ -2,7 +2,7 @@
 
 module Trainees
   class LeadSchoolsController < ApplicationController
-    before_action :authorize_trainee
+    before_action :authorize_trainee, :validate_form_completedness
 
     helper_method :query
 
@@ -25,7 +25,7 @@ module Trainees
       save_strategy = trainee.draft? ? :save! : :stash
 
       if @lead_school_form.public_send(save_strategy)
-        redirect_to origin_page_or_next_step
+        redirect_to step_wizard.next_step
       else
         @school_search = SchoolSearch.call(query: params[:query], lead_schools_only: true)
         render index_or_edit_page
@@ -34,26 +34,12 @@ module Trainees
 
   private
 
-    def redirect_url
-      trainee.requires_employing_school? ? edit_trainee_employing_schools_path(trainee) : trainee_schools_confirm_path(trainee)
-    end
-
     def trainee
       @trainee ||= Trainee.from_param(params[:trainee_id])
     end
 
     def trainee_params
       params.fetch(:schools_lead_school_form, {}).permit(:lead_school_id, *Schools::LeadSchoolForm::NON_TRAINEE_FIELDS)
-    end
-
-    def origin_page_or_next_step
-      return trainee_schools_confirm_path if user_come_from_confirm_or_trainee_page?
-
-      redirect_url
-    end
-
-    def user_come_from_confirm_or_trainee_page?
-      page_tracker.last_origin_page_path&.include?("schools/confirm") || page_tracker.last_origin_page_path == "/trainees/#{trainee.slug}"
     end
 
     def query
@@ -70,6 +56,24 @@ module Trainees
 
     def authorize_trainee
       authorize(trainee, :requires_schools?)
+    end
+
+    def step_wizard
+      @step_wizard ||= Wizards::Schools::StepWizard.new(trainee: trainee, page_tracker: page_tracker)
+    end
+
+    def validate_form_completedness
+      return if user_came_from_backlink?
+
+      redirect_to step_wizard.start_point if any_form_partially_complete?
+    end
+
+    def user_came_from_backlink?
+      session[:origin_path]&.include?("edit")
+    end
+
+    def any_form_partially_complete?
+      step_wizard.any_form_incomplete?
     end
   end
 end
