--- app/forms/diversities/step_wizard.rb
@@ -1,43 +1,0 @@
-# frozen_string_literal: true
-
-module Diversities
-  class StepWizard
-    include Rails.application.routes.url_helpers
-
-    def initialize(trainee:)
-      @trainee = trainee
-    end
-
-    def next_step
-      if disclosure_form.diversity_disclosed?
-        edit_trainee_diversity_ethnic_group_path(trainee)
-      else
-        trainee.apply_application? ? page_tracker.last_origin_page_path : trainee_diversity_confirm_path(trainee)
-      end
-    end
-
-    def any_form_incomplete?
-      ProgressService.call(
-        validator: Diversities::FormValidator.new(trainee),
-        progress_value: trainee.progress.diversity,
-      ).in_progress_invalid?
-    end
-
-    def start_point
-      return edit_trainee_diversity_disclosure_path(trainee) unless DisclosureForm.new(trainee).valid?
-      return edit_trainee_diversity_ethnic_group_path(trainee) unless EthnicGroupForm.new(trainee).valid?
-      return edit_trainee_diversity_ethnic_background_path(trainee) unless EthnicBackgroundForm.new(trainee).valid?
-      return edit_trainee_diversity_disability_disclosure_path(trainee) unless DisabilityDisclosureForm.new(trainee).valid?
-
-      edit_trainee_diversity_disability_detail_path(trainee)
-    end
-
-  private
-
-    def disclosure_form
-      @disclosure_form ||= Diversities::DisclosureForm.new(trainee)
-    end
-
-    attr_reader :trainee
-  end
-end
